now saves progress on tracks
super quick maffs on this, needs development for proper algorithm to determine when it's appropriate to save the spot (don't want seeking to get stored)

+++ HonYomi/Database/FileProgress.cs
@@ -10,6 +10,6 @@ namespace DataLib
         public IdentityUser       User           { get; set; }
         public Guid         FileId         { get; set; }
         public IndexedFile File           { get; set; }
-        public uint        Progress       { get; set; } //in seconds
+        public double        Progress       { get; set; } //in seconds
     }
 }
